Updated save() to accept a zonefile array as a parameter rather than using one loaded by the class. This facilitates using the class to parse a file, manipulating the zone externally, then passing themanipulated zone back into the class to be saved as a zonefile.
Future commits will move the parsing functions to File_DNS_Parser
and add functions in File_DNS to manipulate a loaded zone. There
will be add_rr() functions that accept a Net_DNS2_RR object as well
as add_{$type} functions (ex, add_a($lhs, $ttl, $class, $data)) to
add a record of a specific type. Also in the works is updating this
package to be a PHP5 package.